Using size_t instead of int for sizes

+++ include/algorithms/searching/binary_search.h
@@ -1,4 +1,5 @@
 #pragma once
 
-// TODO: remove size
-int searching_binary_search(int array[], int size, int element);
+#include <stddef.h>
+
+int searching_binary_search(int array[], size_t size, int element);
